+++ apps/shared/curve_view_range.cpp
@@ -1,5 +1,5 @@
 #include "curve_view_range.h"
-#include <math.h>
+#include <cmath>
 #include <ion.h>
 #include <assert.h>
 #include <stddef.h>
@@ -31,12 +31,12 @@ float CurveViewRange::computeGridUnit(Axis axis, float min, float max) {
   float units[3] = {k_smallGridUnitMantissa, k_mediumGridUnitMantissa, k_largeGridUnitMantissa};
   for (int k = 0; k < 3; k++) {
     float unit = units[k];
-    if (floorf(log10f(d/(unit*maxNumberOfUnits))) != floorf(log10f(d/(unit*minNumberOfUnits)))) {
-      b = floorf(log10f(d/(unit*minNumberOfUnits)));
+    if (std::floor(std::log10(d/(unit*maxNumberOfUnits))) != std::floor(std::log10(d/(unit*minNumberOfUnits)))) {
+      b = std::floor(std::log10(d/(unit*minNumberOfUnits)));
       a = unit;
     }
   }
-  return a*powf(10,b);
+  return a*std::pow(10.0f,b);
 }
 
 }
