+++ poincare/src/derivative.cpp
@@ -1,10 +1,9 @@
 #include <poincare/derivative.h>
 #include <poincare/symbol.h>
 #include <poincare/complex.h>
-
+#include <cmath>
 extern "C" {
 #include <assert.h>
-#include <math.h>
 #include <float.h>
 }
 
@@ -52,11 +51,11 @@ Evaluation * Derivative::privateEvaluate(Context& context, AngleUnit angleUnit) 
    * pp. 75–76. */
 
   // Initiate hh
-  float h = fabsf(x) < FLT_MIN ? k_minInitialRate : x/1000.0f;
+  float h = std::fabs(x) < FLT_MIN ? k_minInitialRate : x/1000.0f;
   float f2 = approximateDerivate2(x, h, xContext, angleUnit);
-  f2 = fabsf(f2) < FLT_MIN ? k_minInitialRate : f2;
-  float hh = sqrtf(fabsf(functionValue/(f2/(powf(h,2.0f)))))/10.0f;
-  hh = fabsf(hh) <FLT_MIN ? k_minInitialRate : hh;
+  f2 = std::fabs(f2) < FLT_MIN ? k_minInitialRate : f2;
+  float hh = std::sqrt(std::fabs(functionValue/(f2/(std::pow(h,2.0f)))))/10.0f;
+  hh = std::fabs(hh) <FLT_MIN ? k_minInitialRate : hh;
   /* Make hh an exactly representable number */
   volatile float temp =  x+hh;
   hh = temp - x;
@@ -84,7 +83,7 @@ Evaluation * Derivative::privateEvaluate(Context& context, AngleUnit angleUnit) 
     for (int j = 1; j < 10; j++) {
       a[j][i] = (a[j-1][i]*fac-a[j-1][i-1])/(fac-1.0f);
       fac = k_rateStepSize*k_rateStepSize*fac;
-      errt = fabsf(a[j][i]-a[j-1][i]) > fabsf(a[j][i]-a[j-1][i-1]) ? fabsf(a[j][i]-a[j-1][i]) : fabsf(a[j][i]-a[j-1][i-1]);
+      errt = std::fabs(a[j][i]-a[j-1][i]) > std::fabs(a[j][i]-a[j-1][i-1]) ? std::fabs(a[j][i]-a[j-1][i]) : std::fabs(a[j][i]-a[j-1][i-1]);
       /* Update error and answer if error decreases */
       if (errt < err) {
         err = errt;
@@ -93,7 +92,7 @@ Evaluation * Derivative::privateEvaluate(Context& context, AngleUnit angleUnit) 
     }
     /* If higher extrapolation order significantly increases the error, return
      * early */
-    if (fabsf(a[i][i]-a[i-1][i-1]) > 2.0f*err) {
+    if (std::fabs(a[i][i]-a[i-1][i-1]) > 2.0f*err) {
       break;
     }
   }
@@ -104,8 +103,8 @@ Evaluation * Derivative::privateEvaluate(Context& context, AngleUnit angleUnit) 
   if (err < FLT_MIN) {
     return new Complex(Complex::Float(ans));
   }
-  err = powf(10.0f, (int)log10f(fabsf(err))+2.0f);
-  return new Complex(Complex::Float(roundf(ans/err)*err));
+  err = std::pow(10.0f, (int)std::log10(std::fabs(err))+2.0f);
+  return new Complex(Complex::Float(std::round(ans/err)*err));
 }
 
 float Derivative::growthRateAroundAbscissa(float x, float h, VariableContext xContext, AngleUnit angleUnit) const {
