+++ poincare/test/parser.cpp
@@ -1,7 +1,7 @@
 #include <quiz.h>
 #include <poincare.h>
 #include <ion.h>
-#include <math.h>
+#include <cmath>
 #include <assert.h>
 #include "helper.h"
 
@@ -10,13 +10,13 @@ using namespace Poincare;
 QUIZ_CASE(poincare_parser) {
   Complex a[1] = {Complex::Float(1.2f*M_E)};
   assert_parsed_expression_evaluate_to("1.2*X^(1)", a, 1);
-  Complex b[1] = {Complex::Float(powf(M_E, 2.0f)*M_E)};
+  Complex b[1] = {Complex::Float(std::pow((float)M_E, 2.0f)*M_E)};
   assert_parsed_expression_evaluate_to("X^2*X^(1)", b, 1);
-  Complex c[1] = {Complex::Float(2.0f*powf(3.0f, 4.0f)+2.0f)};
+  Complex c[1] = {Complex::Float(2.0f*std::pow(3.0f, 4.0f)+2.0f)};
   assert_parsed_expression_evaluate_to("2*3^4+2", c, 1);
-  Complex d[1] = {Complex::Float(-2.0f*powf(3.0f, 4.0f)+2.0f)};
+  Complex d[1] = {Complex::Float(-2.0f*std::pow(3.0f, 4.0f)+2.0f)};
   assert_parsed_expression_evaluate_to("-2*3^4+2", d,1);
-  Complex e[1] = {Complex::Float(-sinf(3.0f)*2.0f-3.0f)};
+  Complex e[1] = {Complex::Float(-std::sin(3.0f)*2.0f-3.0f)};
   assert_parsed_expression_evaluate_to("-sin(3)*2-3", e, 1, Radian);
   Complex f[1] = {Complex::Float(-0.003f)};
   assert_parsed_expression_evaluate_to("-.003", f, 1);
@@ -28,7 +28,7 @@ QUIZ_CASE(poincare_parser) {
   assert_parsed_expression_evaluate_to("2/3-5", i, 1);
   Complex j[1] = {Complex::Float(-2.0f/3.0f-5.0f)};
   assert_parsed_expression_evaluate_to("-2/3-5", j, 1);
-  Complex k[1] = {Complex::Float(sinf(3.0f)*2.0f*(4.0f+2.0f))};
+  Complex k[1] = {Complex::Float(std::sin(3.0f)*2.0f*(4.0f+2.0f))};
   assert_parsed_expression_evaluate_to("sin(3)2(4+2)", k, 1, Radian);
   Complex l[1] = {Complex::Float(4.0f/2.0f*(2.0f+3.0f))};
   assert_parsed_expression_evaluate_to("4/2*(2+3)", l, 1, Radian);
